refactor: Move lib and oauth2 to common folder
