+++ src/common/lib/oauth.js
