+++ src/common/lib/oauth.test.js
