+++ src/common/lib/scenario-headers.js
