+++ src/common/lib/scenario-headers.test.js
